requirements.txt
proxmoxer
python-dotenv
requests